Android/Backend/requirements.txt
fastapi
uvicorn[standard]
websockets
yt-dlp
pydantic
python-multipart
aiofiles
python-dotenv
